64bit compilability fix
Note: I briefly tested functionality on Linux 2.6.16.9 x86_64
      and it compiles and runs fine (even opening document and manipulating
      work). No rigourous testing was done though ...

+++ src/gui/basegui.cc
@@ -476,7 +476,7 @@ QVariant BaseGUI::mergeDialog() {
    size_t *res_pos=result->getPositions();
    for (int i=0;i<n;i++) {
     rItems.append(res_items[i]);
-    rPos.append(res_pos[i]);
+    rPos.append((int)res_pos[i]);
    }
    QValueList<QVariant> res;
    res.append(rItems);
